Combine test cases to one file.
